LayoutTests/platform/glib/accessibility/aria-tab-roles-expected.txt

Crust
Veges
Select Crust

This tests that the aria roles for tab, tabpanel and tablist work as expected correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


tabList.role = AXRole: AXTabGroup
tab1.role = AXRole: AXTab
tab1.title = AXTitle: Crust
PASS tab1.childrenCount is 0
tab2.role = AXRole: AXTab
tab2.title = AXTitle: Veges
tabPanel.role = AXRole: AXScrollArea
tabPanel.subrole =
PASS successfullyParsed is true

TEST COMPLETE
